+++ anathema-widgets/src/components/mod.rs
@@ -58,11 +58,14 @@ impl ComponentRegistry {
         self.0.insert_at(id, comp_type);
     }
 
-    /// # Panics
-    ///
-    /// Panics if the component isn't registered.
-    /// This shouldn't happen as the statement eval should catch this.
-    pub fn get(&mut self, id: ComponentBlueprintId) -> Option<(ComponentKind, Box<dyn AnyComponent>, Box<dyn State>)> {
+    // # Panics
+    //
+    // Panics if the component isn't registered.
+    // This shouldn't happen as the statement eval should catch this.
+    pub(super) fn get(
+        &mut self,
+        id: ComponentBlueprintId,
+    ) -> Option<(ComponentKind, Box<dyn AnyComponent>, Box<dyn State>)> {
         match self.0.get_mut(id) {
             Some(component) => match component {
                 ComponentType::Component(comp, state) => Some((ComponentKind::Instance, comp.take()?, state.take()?)),
@@ -117,17 +120,23 @@ impl<T> Copy for ComponentId<T> {}
 
 pub struct ViewMessage {
     pub(super) payload: Box<dyn Any + Send + Sync>,
-    pub(super) recipient: ComponentBlueprintId,
+    pub(super) recipient: Recipient,
 }
 
 impl ViewMessage {
-    pub fn recipient(&self) -> ComponentBlueprintId {
+    pub fn recipient(&self) -> Recipient {
         self.recipient
     }
 
     pub fn payload(self) -> Box<dyn Any + Send + Sync> {
         self.payload
     }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub enum Recipient {
+    ComponentId(ComponentBlueprintId),
+    WidgetId(WidgetId),
 }
 
 #[derive(Debug, Clone)]
@@ -139,7 +148,15 @@ impl From<flume::Sender<ViewMessage>> for Emitter {
     }
 }
 
+pub type MessageReceiver = flume::Receiver<ViewMessage>;
+
 impl Emitter {
+    pub fn new() -> (Self, MessageReceiver) {
+        let (tx, rx) = flume::unbounded();
+        (Self(tx), rx)
+    }
+
+    /// Emit a message to a component via a `ComponentId<T>`
     pub fn emit<T: 'static + Send + Sync>(
         &self,
         component_id: ComponentId<T>,
@@ -147,11 +164,12 @@ impl Emitter {
     ) -> Result<(), SendError<ViewMessage>> {
         let msg = ViewMessage {
             payload: Box::new(value),
-            recipient: component_id.0,
+            recipient: Recipient::ComponentId(component_id.0),
         };
         self.0.send(msg)
     }
 
+    /// Emit a message to a component via a `ComponentId<T>` asynchronously
     pub async fn emit_async<T: 'static + Send + Sync>(
         &self,
         component_id: ComponentId<T>,
@@ -159,7 +177,35 @@ impl Emitter {
     ) -> Result<(), SendError<ViewMessage>> {
         let msg = ViewMessage {
             payload: Box::new(value),
-            recipient: component_id.0,
+            recipient: Recipient::ComponentId(component_id.0),
+        };
+        self.0.send_async(msg).await
+    }
+
+    /// Try to emit a message to a component via its `WidgetId`.
+    /// Unlike the `emit` function there is no type associated with the widget id,
+    /// and the message will fail to deliver silently if the wrong type is used.
+    pub fn try_emit<T: 'static + Send + Sync>(
+        &self,
+        widget_id: WidgetId,
+        value: T,
+    ) -> Result<(), SendError<ViewMessage>> {
+        let msg = ViewMessage {
+            payload: Box::new(value),
+            recipient: Recipient::WidgetId(widget_id),
+        };
+        self.0.send(msg)
+    }
+
+    /// Async counterpart of `try_emit`.
+    pub async fn try_emit_async<T: 'static + Send + Sync>(
+        &self,
+        widget_id: WidgetId,
+        value: T,
+    ) -> Result<(), SendError<ViewMessage>> {
+        let msg = ViewMessage {
+            payload: Box::new(value),
+            recipient: Recipient::WidgetId(widget_id),
         };
         self.0.send_async(msg).await
     }
@@ -196,7 +242,7 @@ impl<'frame, 'bp, T: 'static> Context<'frame, 'bp, T> {
             parent,
             *assoc_event_map,
             self.ident_id,
-            self.sender_id,
+            self.widget_id,
             data,
         );
     }
@@ -231,7 +277,7 @@ impl<'frame, 'bp, T> DerefMut for Context<'frame, 'bp, T> {
 pub struct AnyComponentContext<'frame, 'bp> {
     parent: Option<Parent>,
     ident_id: StringId,
-    sender_id: WidgetId,
+    pub widget_id: WidgetId,
     state_id: StateId,
     assoc_functions: &'frame [AssocEventMapping],
     assoc_events: &'frame mut AssociatedEvents,
@@ -263,7 +309,7 @@ impl<'frame, 'bp> AnyComponentContext<'frame, 'bp> {
         Self {
             parent,
             ident_id,
-            sender_id,
+            widget_id: sender_id,
             state_id,
             assoc_functions,
             assoc_events,